Добавлены диаграммы активности и компонентов.
