+++ GUI/Scripting/CmdReport.cs
@@ -1,10 +1,8 @@
-﻿using BazisGUI.Scripting.Variable;
-
-namespace BazisGUI.Scripting
+﻿namespace BazisGUI.Scripting
 {
     public class CmdReport
     {
         public string Name { get; set; }
-        public ScriptValue Variable {  get; set; }
+        public string Variable {  get; set; }
     }
 }
