--- GUI/Scripting/Variable/IntValue.cs
@@ -1,9 +1,0 @@
-﻿namespace BazisGUI.Scripting.Variable
-{
-    public class IntValue : ScriptValue
-    {
-        public int Value { get; set; }
-
-        public override string ToString() => Value.ToString();
-    }
-}
