--- GUI/Scripting/Variable/ScriptValue.cs
@@ -1,9 +1,0 @@
-﻿using System;
-
-namespace BazisGUI.Scripting.Variable
-{
-    public abstract class ScriptValue
-    {
-        public abstract string ToString();
-    }
-}
