--- GUI/Scripting/Variable/StringValue.cs
@@ -1,9 +1,0 @@
-﻿namespace BazisGUI.Scripting.Variable
-{
-    public class StringValue : ScriptValue
-    {
-        public string Value { get; set; }
-
-        public override string ToString() => Value;
-    }
-}
